requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
beautifulsoup4==4.12.3
    # via -r requirements.in
black==24.10.0
    # via -r requirements.in
click==8.1.8
    # via black
flake8==7.1.1
    # via -r requirements.in
jinja2==3.1.5
    # via -r requirements.in
markdown==3.7
    # via -r requirements.in
markupsafe==3.0.2
    # via jinja2
mccabe==0.7.0
    # via flake8
mypy-extensions==1.0.0
    # via black
packaging==24.2
    # via black
pathspec==0.12.1
    # via black
pillow==11.1.0
    # via -r requirements.in
platformdirs==4.3.6
    # via black
pycodestyle==2.12.1
    # via flake8
pyflakes==3.2.0
    # via flake8
pyyaml==6.0.2
    # via -r requirements.in
smartypants==2.0.1
    # via typogrify
soupsieve==2.6
    # via beautifulsoup4
typogrify==2.1.0
    # via -r requirements.in
watchdog==6.0.0
    # via -r requirements.in
